Move entity tests to one package
