+++ collect_app/src/androidTest/java/org/odk/collect/android/feature/formentry/entities/EntityFormApprovalTest.kt
@@ -1,4 +1,4 @@
-package org.odk.collect.android.feature.formentry
+package org.odk.collect.android.feature.formentry.entities
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Rule
